fix(gateway): 优化用户ID添加逻辑，提升安全性
- 移除对Redis的依赖，改为使用Sa-Token框架直接验证Token并获取用户ID
- 过滤请求头中可能存在的用户ID，防止客户端伪造
- 设置过滤器顺序，提高在网关中优先处理用户ID
- 修改日志，调整敏感信息输出，避免暴露完整手机号
- 修正阿里云OSS文件上传逻辑，改为流式上传防止内存溢出
- 增加阿里云OSS文件删除功能
- 新增聚合批次标识字段，支持2DB幂等键区分不同批次聚合结果
- 优化操作日志序列化，防止MultipartFile和Servlet请求等对象序列化异常
- 修正Vue3搜索组件样式与交互逻辑，提升移动端体验
- 精简Vue3侧边栏导航和隐藏无用路由链接
- 更新fishhub-gateway和fishhub-oss配置，调整超时时间和文件上传大小限制
- 删除fishhub-search无用模块定义及配置文件

+++ fishhub-data-align/src/main/java/hk/ljx/fishhub/data/align/domain/mapper/UpdateMapper.java
@@ -13,41 +13,41 @@ public interface UpdateMapper {
      * @return
      */
     int updateUserFollowingTotalByUserId(@Param("userId") long userId,
-                                         @Param("followingTotal") int followingTotal);
+                                         @Param("followingTotal") long followingTotal);
 
     /**
      * 更新 t_user_count 计数表总粉丝数
      */
     int updateUserFansTotalByUserId(@Param("userId") long userId,
-                                         @Param("fansTotal") int fansTotal);
+                                         @Param("fansTotal") long fansTotal);
 
     /**
      * 更新 t_user_count 计数表获得的总点赞数
      */
     int updateUserLikeTotalByUserId(@Param("userId") long userId,
-                                    @Param("likeTotal") int likeTotal);
+                                    @Param("likeTotal") long likeTotal);
 
     /**
      * 更新 t_user_count 计数表获得的总收藏数
      */
     int updateUserCollectTotalByUserId(@Param("userId") long userId,
-                                    @Param("collectTotal") int collectTotal);
+                                    @Param("collectTotal") long collectTotal);
 
     /**
      * 更新 t_user_count 计数表获得的总笔记发布数
      */
     int updateUserNoteTotalByUserId(@Param("userId") long userId,
-                                       @Param("noteTotal") int noteTotal);
+                                       @Param("noteTotal") long noteTotal);
 
     /**
      * 更新 t_note_count 计数表笔记点赞数
      */
     int updateNoteLikeTotalByUserId(@Param("noteId") long noteId,
-                                         @Param("noteLikeTotal") int noteLikeTotal);
+                                         @Param("noteLikeTotal") long noteLikeTotal);
 
     /**
      * 更新 t_note_count 计数表笔记收藏数
      */
     int updateNoteCollectTotalByUserId(@Param("noteId") long noteId,
-                                    @Param("noteCollectTotal") int noteCollectTotal);
+                                    @Param("noteCollectTotal") long noteCollectTotal);
 }
